requirements.txt
PyQt6>=6.5
pyaudiowpatch>=0.2.12
deepgram-sdk>=3.0
anthropic>=0.30
numpy
